+++ game_patch/rf/player/player.h
@@ -84,6 +84,10 @@ struct PlayerAdditionalData {
     // `std::nullptr` represents freelook spectate mode.
     std::optional<rf::Player*> spectatee{};
     bool remote_server_cfg_sent = false;
+
+    // Server side rail gun reload cooldown, used for force_rail_reload
+    // needed because entity_is_reloading is unreliable on server
+    rf::Timestamp rail_gun_reload_timer{};
 };
 static_assert(alignof(PlayerAdditionalData) == 0x8);
 #endif